[fix] main.goの位置を修正
